moved constant to legacy to avoid changes
