+++ diktat-rules/src/main/kotlin/com/saveourtool/diktat/common/config/rules/LegacyUtils.kt
@@ -8,7 +8,13 @@ import com.saveourtool.diktat.api.DiktatRuleConfig
 import com.saveourtool.diktat.api.DiktatRuleNameAware
 import com.saveourtool.diktat.api.findByRuleName
 
+/**
+ * Name of common configuration
+ */
+const val DIKTAT_COMMON = "DIKTAT_COMMON"
+
 typealias RuleConfiguration = com.saveourtool.diktat.ruleset.config.RuleConfiguration
+typealias CommonConfiguration = com.saveourtool.diktat.ruleset.config.CommonConfiguration
 
 /**
  * Get [DiktatRuleConfig] for particular [DiktatRuleNameAware] object.
@@ -17,3 +23,13 @@ typealias RuleConfiguration = com.saveourtool.diktat.ruleset.config.RuleConfigur
  * @return [DiktatRuleConfig] for a particular rule if it is found, else null
  */
 fun List<DiktatRuleConfig>.getRuleConfig(rule: DiktatRuleNameAware): DiktatRuleConfig? = this.findByRuleName(rule)
+
+/**
+ * @return common configuration from list of all rules configuration
+ */
+fun List<DiktatRuleConfig>.getCommonConfiguration(): CommonConfiguration = CommonConfiguration(getCommonConfig()?.configuration)
+
+/**
+ * Get [DiktatRuleConfig] representing common configuration part that can be used in any rule
+ */
+private fun List<DiktatRuleConfig>.getCommonConfig() = find { it.name == DIKTAT_COMMON }
